drivers: misc: adi-axi-tdd-ng: Add new TDD engine driver
This patch introduces the driver for the new ADI TDD engine HDL,
which aims to fix many of the issues of the old design (See cf_axi_tdd).

The interface is sysfs based and follows the IIO specification, even
tough the device is not actually registered as an IIO device, and
requires a translation layer like iio_fakedev or a modified iiod to be
visible to an iio client.

Signed-off-by: David Winter <david.winter@analog.com>